Replace "(unsigned long) -1L" with -1UL
* defs.h (dumpiov): Replace "(unsigned long) -1L" with -1UL.
* io.c (print_iovec, tprint_iov): Likewise.
* msghdr.c (SYS_FUNC(sendmsg)): Likewise.
* syscall.c (dumpio): Likewise.
* poll.c (decode_poll_exiting): Replace "(unsigned int) -1" with -1U.

+++ defs.h
@@ -592,7 +592,7 @@ extern void dumpiov_in_msghdr(struct tcb *, long, unsigned long);
 extern void dumpiov_in_mmsghdr(struct tcb *, long);
 extern void dumpiov_upto(struct tcb *, int, long, unsigned long);
 #define dumpiov(tcp, len, addr) \
-	dumpiov_upto((tcp), (len), (addr), (unsigned long) -1L)
+	dumpiov_upto((tcp), (len), (addr), -1UL)
 extern void dumpstr(struct tcb *, long, int);
 extern void printstr_ex(struct tcb *, long addr, long len,
 	unsigned int user_style);
